docxmodifier/requirements.txt
django==1.7.0
djangorestframework==3.1.1
django-rest-swagger==0.3.4
python-docx==0.8.6
